Cambios generales

+++ trabajo/src/main/java/com/proyecto/trabajo/dto/UsuariosCreateDto.java
@@ -11,7 +11,7 @@ public class UsuariosCreateDto {
     private String nom_su;
     private String ape_su;
     private String corre;
-    private Integer num_docu;
+    private Long num_docu;
     private String pasword;
     private Byte estad;
